update: 1.增加全局异常处理类; 2.修改申请模块相关接口及处理流程

+++ service-topic/src/main/java/com/dataint/topic/db/dao/IApplicationDao.java
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface IApplicationDao extends JpaRepository<Application, Long> {
 
-    List<Application> findAllByTopicIdAndStatus(Integer topicId, Integer status);
+    List<Application> findAllByTopicIdAndStatus(Long topicId, Integer status);
 
     Page<Application> findAllByStatusAndTopicNameContainingOrderByCreatedTimeDesc(Integer status, String keyword, Pageable pageable);
